don't delete WPA until retired
