+++ pkg/controller/releasemanager/incarnation.go
@@ -457,8 +457,14 @@ func (i *Incarnation) genScalePlan(ctx context.Context) *rmplan.ScaleRevision {
 	max := i.divideReplicas(i.target().Scale.Max)
 	if i.status.CurrentPercent == 0 {
 		min = max
-	} else if i.target().Scale.Worker != nil && *i.target().Scale.Min == 0 {
-		min = 0
+	}
+	if i.target().Scale.Worker != nil {
+		if !i.isRoutable() {
+			min = 0
+			max = 0
+		} else if *i.target().Scale.Min == 0 {
+			min = 0
+		}
 	}
 
 	return &rmplan.ScaleRevision{
